fix: canonicalize receipt side tables
