+++ packages/bijou/src/core/packed-bijou-cells-validate-glyphs.ts
@@ -1,14 +1,12 @@
-import {
-  graphemeClusterWidth,
-  sanitizePlainTerminalText,
-  segmentGraphemes,
-} from './text/grapheme.js';
 import { SIDE_TABLE_THRESHOLD } from './render/packed-cell.js';
 import {
-  MAX_PACKED_BIJOU_GLYPH_CODE_UNITS,
   MAX_PACKED_BIJOU_SIDE_TABLE_ENTRIES,
   PACKED_BIJOU_CELL_STRIDE,
 } from './packed-bijou-cells-contract.js';
+import {
+  isCanonicalSideTableGlyph,
+  isSafeUnitGlyph,
+} from './packed-bijou-cells-glyph-policy.js';
 import {
   arrayAt,
   failPackedCells,
@@ -27,7 +25,7 @@ export function validateSideTable(
   const seen = new Set<string>();
   const sideTable = raw.map((value, index) => {
     const path = `$.sideTable[${String(index)}]`;
-    if (typeof value !== 'string' || !isCanonicalUnitGlyph(value)) {
+    if (typeof value !== 'string' || !isCanonicalSideTableGlyph(value)) {
       failPackedCells(
         'invalid-glyph',
         path,
@@ -81,7 +79,7 @@ function validateCellGlyphs(
         `$.bytes[${String(offset)}]`,
         'direct glyph cannot be a UTF-16 surrogate',
       );
-    } else if (code !== 0 && !isCanonicalUnitGlyph(String.fromCharCode(code))) {
+    } else if (code !== 0 && !isSafeUnitGlyph(String.fromCharCode(code))) {
       failPackedCells(
         'invalid-glyph',
         `$.bytes[${String(offset)}]`,
@@ -90,50 +88,6 @@ function validateCellGlyphs(
     }
   }
   return references;
-}
-
-function isCanonicalUnitGlyph(value: string): boolean {
-  return (
-    value.length > 0 &&
-    value.length <= MAX_PACKED_BIJOU_GLYPH_CODE_UNITS &&
-    !hasTerminalControl(value) &&
-    !hasUnsafeFormat(value) &&
-    hasVisibleBase(value) &&
-    sanitizePlainTerminalText(value) === value &&
-    segmentGraphemes(value).length === 1 &&
-    graphemeClusterWidth(value) === 1
-  );
-}
-
-function hasVisibleBase(value: string): boolean {
-  return Array.from(value).some(
-    (character) => !/^[\p{M}\p{Cf}]$/u.test(character),
-  );
-}
-
-function hasUnsafeFormat(value: string): boolean {
-  return Array.from(value).some((character) => {
-    const codePoint = character.codePointAt(0);
-    if (codePoint === 0x200c || codePoint === 0x200d) return false;
-    if (
-      codePoint !== undefined &&
-      codePoint >= 0xe0020 &&
-      codePoint <= 0xe007f
-    ) {
-      return false;
-    }
-    return /^\p{Cf}$/u.test(character);
-  });
-}
-
-function hasTerminalControl(value: string): boolean {
-  return Array.from(value).some((character) => {
-    const codePoint = character.codePointAt(0);
-    return (
-      codePoint !== undefined &&
-      (codePoint <= 0x1f || (codePoint >= 0x7f && codePoint <= 0x9f))
-    );
-  });
 }
 
 function byte(bytes: readonly number[], index: number): number {
